Handle optional Vec and BTreeMap as Option<t>
Previously, optional array properties are required in the generated struct, but skipped if serializing. This is incorrect behavior, since an empty Vec is not the same as not supplying a Vec.

This breaks server-side apply-ing new resources to remove things from the Vec, as well as cases where a property is required in a one-of, but you actually want to send the empty Vec as one of those properties.

This change instead makes it an Option<Vec<T>> with #[serde(default, skip_serializing_if = "Option::is_none")].

The above also applies to BTreeMaps.

Resolves https://github.com/kube-rs/kopium/issues/40

Signed-off-by: Alex Hunt <alex.hunt@materialize.com>