Add optional snake_case conversion of member names
Using heck as it's the most popular libary out there (plus it's
maintained by boats).

This CAN lead to some subtle behaviour due to how serde's rename_all is
not a true inverse of ToSnakeCase, but I believe this can be rectified
for those cases with some extra override flags (awkward tho it may be).

fixes #22

Signed-off-by: clux <sszynrae@gmail.com>