add support for standard Condition APIs
Detect the presence of a Conditions object in a schema and use
`k8s_openapi::apimachinery::pkg::apis::meta::v1::Condition` if present.
Introduce a new flag `--no-condition` which allows revering back to the
old behavior of generating a custom Condition API for each instance.

Furthermore, introduce `analyzer::Config` which allows for customizing
the behavior of the `analyze` function.

Signed-off-by: Sanskar Jaiswal <jaiswalsanskar078@gmail.com>
Co-authored-by: clux <sszynrae@gmail.com>