Add relaxed mode for partially invalid CRDs
Fixes #108

Argo workflows have 2 CRDs, minimal and full.
The full CRD is supposed to be used in the editor and can't be
installed in kubernetes, but provide full validation.
The minimal version has no validation and should be installed.

In `--relaxed` mode, entries like `inline: {}` are mapped to
object like hashmaps.

Signed-off-by: Daniel Poelzleithner <poelzleithner@b1-systems.de>